Refactor the email validation
The alert message  was not read by Narrator, so I made some changes:
1. Remove the "aria-live" attribute from the alert element
2. Set the "role" as "alert" dynamically through JavaScript

I also add new function to remove the alert message, so that when the
user try again and still input the invalid email, then the screen reader
will pronounce the alert message again.

So, with all the changes, I also make all the DOM variables as global
variables.